backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
opencv-python==4.9.0.80
Pillow==10.2.0
werkzeug==3.0.1
